# Conflicts:
#	.idea/workspace.xml

+++ patient-service/src/main/java/org/fpic974/patientservice/service/PatientService.java
@@ -9,6 +9,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -19,10 +20,11 @@ public class PatientService implements IPatientService {
 
     @Override
     public List<PatientResponse> getAllPatients() {
-        log.debug("Service Call : getAllPatients()");
+        log.debug(">> Calling method : getAllPatients()");
+
         List<Patient> patients = patientRepository.findAll();
 
-        log.debug("Result : " + patients);
+        log.debug(">> Result         : {} object(s)", patients.size());
 
         return patients.stream()
                 .map(patient -> PatientResponse.builder()
@@ -38,27 +40,33 @@ public class PatientService implements IPatientService {
 
     @Override
     public PatientResponse getPatientById(Integer id) {
-        log.debug("Service Call : getPatientById({})", id);
+        log.debug(">> Calling method : getPatientById({})", id);
 
-        Patient patient = patientRepository.findById(id)
-                .orElseThrow();
+        Optional<Patient> patient = patientRepository.findById(id);
 
-        log.debug("Result : " + patient);
+        if (patient.isEmpty()) {
+            log.debug("<< No object found with id {}", id);
+            throw new IllegalArgumentException("Invalid id");
+        }
+
+        Patient patientFound = patient.get();
+
+        log.debug("<< Result         : " + patientFound.getId());
 
         return PatientResponse.builder()
-                .id(patient.getId())
-                .lastName(patient.getLastName())
-                .firstName(patient.getFirstName())
-                .birthDate(patient.getBirthDate())
-                .gender(patient.getGender())
-                .address(patient.getAddress())
-                .phone(patient.getPhone())
+                .id(patientFound.getId())
+                .lastName(patientFound.getLastName())
+                .firstName(patientFound.getFirstName())
+                .birthDate(patientFound.getBirthDate())
+                .gender(patientFound.getGender())
+                .address(patientFound.getAddress())
+                .phone(patientFound.getPhone())
                 .build();
     }
 
     @Override
     public PatientResponse createPatient(PatientRequest patientRequest) {
-        log.debug("Service Call : createPatient({})", patientRequest);
+        log.debug(">> Calling method : createPatient({})", patientRequest);
 
         Patient patient = Patient.builder()
                 .lastName(patientRequest.getLastName())
@@ -69,33 +77,43 @@ public class PatientService implements IPatientService {
                 .phone(patientRequest.getPhone())
                 .build();
 
-        patientRepository.save(patient);
+        Patient patientSaved = patientRepository.save(patient);
 
-        log.debug("Result : Patient with id {} created", patient.getId());
+        log.debug("<< Result         : object with id {} created", patient.getId());
 
         return PatientResponse.builder()
-                .id(patient.getId())
-                .lastName(patient.getLastName())
-                .firstName(patient.getFirstName())
-                .birthDate(patient.getBirthDate())
-                .gender(patient.getGender())
-                .address(patient.getAddress())
-                .phone(patient.getPhone())
+                .id(patientSaved.getId())
+                .lastName(patientSaved.getLastName())
+                .firstName(patientSaved.getFirstName())
+                .birthDate(patientSaved.getBirthDate())
+                .gender(patientSaved.getGender())
+                .address(patientSaved.getAddress())
+                .phone(patientSaved.getPhone())
                 .build();
     }
 
     @Override
     public void deletePatientById(Integer id) {
-        log.debug("Service Call : deletePatientById({})", id);
+        log.debug(">> Calling method : deletePatientById({})", id);
+
+        if (!patientRepository.existsById(id)) {
+            log.debug("<< No object found with id {}", id);
+            throw new IllegalArgumentException("Invalid id");
+        }
 
         patientRepository.deleteById(id);
 
-        log.debug("Result : Patient with id {} deleted", id);
+        log.debug("<< Result         : object with id {} deleted", id);
     }
 
     @Override
     public PatientResponse updatePatientById(Integer id, PatientRequest patientRequest) {
-        log.debug("Service Call : updatePatientById({}, {})", id, patientRequest);
+        log.debug(">> Calling method : updatePatientById({}, {})", id, patientRequest);
+
+        if (!patientRepository.existsById(id)) {
+            log.debug("<< No object found with id {}", id);
+            throw new IllegalArgumentException("Invalid id");
+        }
 
         Patient patient = Patient.builder()
                 .id(id)
@@ -109,7 +127,7 @@ public class PatientService implements IPatientService {
 
         patientRepository.save(patient);
 
-        log.debug("Result : Patient with id {} updated", patient.getId());
+        log.debug(">> Result         : object with id {} updated", patient.getId());
 
         return PatientResponse.builder()
                 .id(patient.getId())
